requirements.txt
apscheduler
psycopg2-binary
python-dotenv
pytz
requests
sqlalchemy